# Avoid cloning a TtToken in SubtreeTokenSource::mk_token