@lnicola 2019-08-03
CC https://github.com/rust-analyzer/rowan/blob/master/src/green.rs#L150