@kjeremy 2019-08-03
The number of things processed look different between those two runs.