@lnicola 2019-08-03
> The number of things processed look different between those two runs.

If you mean "Total expressions", I think I changed the code in the meanwhile.